Fix warning about pointer/int cast
git-svn-id: svn://svn.code.sf.net/p/x10/code@10728 a326200e-df2c-42e0-81fa-597e909af41d

+++ trunk/x10.runtime.17/src-cpp/x10aux/ref.cc
@@ -27,7 +27,7 @@ remote_ref remote_ref::make (void *ptr) {
 }
 
 void *remote_ref::take (remote_ref r) {
-    if (r.loc==x10aux::here()) return (void*)r.addr;
+    if (r.loc==x10aux::here()) return (void*)(size_t)r.addr;
     if (r.addr==0) return NULL;
     return mask(new remote_ref(r));
 }
